+++ housecut/app/src/main/java/com/housecut/housecut/household_member_class/makefile.mak
